+++ Core/src/Core/PipelineContext.cs
@@ -1,7 +1,7 @@
 ﻿// <copyright file="PipelineContext.cs" company="Chris Trout">
 // MIT License
 //
-// Copyright(c) 2019-2022 Chris Trout
+// Copyright(c) 2019-2023 Chris Trout
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -27,6 +27,7 @@ namespace MyTrout.Pipelines.Core
     using Microsoft.Extensions.Configuration;
     using System;
     using System.Collections.Generic;
+    using System.Reflection;
     using System.Threading;
 
     /// <summary>
@@ -34,9 +35,17 @@ namespace MyTrout.Pipelines.Core
     /// </summary>
     public class PipelineContext : IPipelineContext
     {
+        /*
+         * The PipelineName property has a default value that cannot be complete covered by unit tests.
+         * There is no realistic way to test returning null from GetEntryAssembly() or the Name property.
+         * Because this property's initialization is executed in the constructor in code, the code coverage
+         * of AltCover instruments this as a constructor failure rather than a property failure.
+         */
+
         /// <summary>
         /// Initializes a new instance of the <see cref="PipelineContext" /> class.
         /// </summary>
+        [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
         public PipelineContext()
         {
             // no op
@@ -73,6 +82,12 @@ namespace MyTrout.Pipelines.Core
         /// Gets context items which are passed between step instances during execution.
         /// </summary>
         public IDictionary<string, object> Items { get; } = new Dictionary<string, object>();
+
+        /// <summary>
+        /// Gets the name of the Pipeline to be used for logging.
+        /// </summary>
+        /// <remarks>If the name cannot be read from the EntryAssembly, the value is defaulted to "Name Not Available".</remarks>
+        public string PipelineName { get; init; } = Assembly.GetEntryAssembly()?.GetName().Name ?? "Name Not Available";
 
         /// <summary>
         /// Gets a UTC timestamp representing when this pipeline started its execution.