+++ Hosting/src/PipelineHostedService.cs
@@ -1,7 +1,7 @@
 ﻿// <copyright file="PipelineHostedService.cs" company="Chris Trout">
 // MIT License
 //
-// Copyright(c) 2019-2020 Chris Trout
+// Copyright(c) 2019-2022 Chris Trout
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -35,34 +35,27 @@ namespace MyTrout.Pipelines.Hosting
     /// <summary>
     /// Provides a service to run pipelines using the <see cref="IHost" />.
     /// </summary>
-    public class PipelineHostedService : IHostedService
+    public class PipelineHostedService : BackgroundService
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="PipelineHostedService" /> class with the specified parameters.
         /// </summary>
         /// <param name="logger">The logger for the pipeline application service.</param>
-        /// <param name="applicationLifetime">The lifetime of the <see cref="IHost" /> application.</param>
         /// <param name="stepActivator">The activator that constructs steps from <see cref="Type" />.</param>
         /// <param name="pipelineBuilder">The <see cref="PipelineBuilder" /> hosted by this <see cref="IHost" />.</param>
         /// <param name="context">The <see cref="PipelineContext" /> used by this <see cref="PipelineContext" />.</param>
         public PipelineHostedService(
                                 ILogger<PipelineHostedService> logger,
-                                IHostApplicationLifetime applicationLifetime,
                                 IStepActivator stepActivator,
                                 PipelineBuilder pipelineBuilder,
                                 PipelineContext context)
+            : base()
         {
-            this.ApplicationLifetime = applicationLifetime ?? throw new ArgumentNullException(nameof(applicationLifetime));
             this.Context = context ?? throw new ArgumentNullException(nameof(context));
             this.Logger = logger ?? throw new ArgumentNullException(nameof(logger));
             this.PipelineBuilder = pipelineBuilder ?? throw new ArgumentNullException(nameof(pipelineBuilder));
             this.StepActivator = stepActivator ?? throw new ArgumentNullException(nameof(stepActivator));
         }
-
-        /// <summary>
-        /// Gets the lifetime of the hosted application.
-        /// </summary>
-        public IHostApplicationLifetime ApplicationLifetime { get; }
 
         /// <summary>
         /// Gets the logger for this pipeline application service.
@@ -84,59 +77,50 @@ namespace MyTrout.Pipelines.Hosting
         /// </summary>
         public IStepActivator StepActivator { get; }
 
-        /* NOTES TO DEVELOPERS: The original implementation below used an async lambda until the Task.RunSynchronously() blocking
-         *                      thread method was found.  This stupidity is required because GenericHost implementors
-         *                      did not account for an async ApplicationStarted delegate and how exceptions could be thrown
-         *                      from a void delegate returning from the ApplicationStarted capability which would need to be
-         *                      handled via an AppDomain.Current.UnhandledException event handler.  The implementation below does
-         *                      not suffer from the same issue for **MOST** exceptions at the cost of blocking threads on a delegate call.
-         */
-
         /// <inheritdoc />
-        public Task StartAsync(CancellationToken cancellationToken)
+        public override Task StartAsync(CancellationToken cancellationToken)
         {
             this.Logger.LogDebug(Resources.CALLING_METHOD_IN_TYPE(CultureInfo.CurrentCulture, nameof(this.StartAsync), nameof(PipelineHostedService)));
 
-            this.ApplicationLifetime.ApplicationStarted.Register(() => this.OnApplicationStarted(this.ApplicationLifetime.ApplicationStopping));
-            return Task.CompletedTask;
+            return base.StartAsync(cancellationToken);
         }
 
         /// <inheritdoc />
-        public Task StopAsync(CancellationToken cancellationToken)
+        public override Task StopAsync(CancellationToken cancellationToken)
         {
             this.Logger.LogDebug(Resources.CALLING_METHOD_IN_TYPE(CultureInfo.CurrentCulture, nameof(this.StopAsync), nameof(PipelineHostedService)));
-            return Task.CompletedTask;
+            return base.StopAsync(cancellationToken);
         }
+#pragma warning disable S125
+        /*
+         * NOTE TO FUTURE DEVELOPERS: According to CS1058, all non-CLS exceptions will be wrapped in a RuntimeWrappedException
+         *                            unless a specific assembly level attribute is configured.
+         *                            RuntimeCompatibilityAttribute(WrapNonExceptionThrows = false)]
+         *                            DO NOT CONFIGURE THIS ATTRIBUTE unless you add a catch() statement here or exceptions could be unhandled.
+        */
+#pragma warning restore S125
 
-        /// <summary>
-        /// Forces the OnApplicationStartedAsync to be run synchronously and block the calling thread.
-        /// </summary>
-        /// <param name="cancellationToken">Indicates that the application execution has been aborted.</param>
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Usage", "VSTHRD002:Avoid problematic synchronous waits", Justification = "Using GetResult() in Console Applications where multiple threads are not being run is acceptable.")]
-
-        private void OnApplicationStarted(CancellationToken cancellationToken)
+        protected override async Task ExecuteAsync(CancellationToken stoppingToken)
         {
-            this.OnApplicationStartedAsync(cancellationToken).GetAwaiter().GetResult();
-        }
-
-        /// <summary>
-        /// Executes the configured pipelines when the application is started.
-        /// </summary>
-        /// <param name="cancellationToken">Indicates that the application execution has been aborted.</param>
-        /// <returns>A <see cref="Task" />.</returns>
-        private async Task OnApplicationStartedAsync(CancellationToken cancellationToken)
-        {
-            this.Context.CancellationToken = cancellationToken;
-
             try
             {
-                this.Logger.LogInformation(Resources.PIPELINE_TASKS_STARTING(CultureInfo.CurrentCulture));
+                this.Context.CancellationToken = stoppingToken;
+
+                this.Logger.LogInformation(Resources.PIPELINE_TASKS_STARTING(CultureInfo.CurrentCulture, this.Context.PipelineName));
+
+                DateTime startingPipelineBuild = DateTime.UtcNow;
 
                 IPipelineRequest pipeline = this.PipelineBuilder.Build(this.StepActivator);
 
+                this.Logger.LogDebug(Resources.PIPELINE_BUILD_ELAPSED_TIME(CultureInfo.CurrentCulture), this.Context.PipelineName, (DateTime.UtcNow - startingPipelineBuild).TotalSeconds);
+
+                DateTime startingPipelineExecution = DateTime.UtcNow;
+
                 await pipeline.InvokeAsync(this.Context).ConfigureAwait(false);
 
-                this.Logger.LogInformation(Resources.TASKS_ARE_COMPLETED(CultureInfo.CurrentCulture));
+                this.Logger.LogDebug(Resources.PIPELINE_EXECUTION_ELAPSED_TIME(CultureInfo.CurrentCulture), this.Context.PipelineName, (DateTime.UtcNow - startingPipelineExecution).TotalSeconds);
+
+                this.Logger.LogInformation(Resources.TASKS_ARE_COMPLETED(CultureInfo.CurrentCulture, this.Context.PipelineName));
             }
             catch (Exception exc)
             {
@@ -146,10 +130,10 @@ namespace MyTrout.Pipelines.Hosting
             {
                 foreach (var error in this.Context.Errors)
                 {
-                    this.Logger.LogError(error, Resources.HANDLED_EXCEPTION(CultureInfo.CurrentCulture));
+                    this.Logger.LogError(error, Resources.HANDLED_EXCEPTION(CultureInfo.CurrentCulture, this.Context.PipelineName));
                 }
 
-                this.ApplicationLifetime.StopApplication();
+                this.Logger.LogDebug(Resources.PIPELINE_TOTAL_EXECUTION_ELAPSED_TIME(CultureInfo.CurrentCulture, this.Context.PipelineName), (DateTime.UtcNow - this.Context.Timestamp).TotalSeconds);
             }
         }
     }